+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simple connection management that selects the most recent connection for store, light push and filter requests.
 - `createLightNode` to create a Waku node for resource restricted environment with Light Push, Filter and Relay.
 - `createPrivacyNode` to create a Waku node for privacy preserving usage with Relay only.
+- `createFullNode` to create a Waku node for with all protocols, for **testing purposes only**.
 
 ### Changed